move token string handling to tokenstr module
