+++ relengapi/blueprints/tokenauth/tokenstr.py
@@ -1,0 +1,38 @@
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+import logging
+
+from flask import current_app
+from itsdangerous import BadData
+from itsdangerous import JSONWebSignatureSerializer
+
+TOKENAUTH_VERSION = 1
+logger = logging.getLogger(__name__)
+
+
+def init_app(app):
+    if not app.secret_key:
+        logger.warning("The `SECRET_KEY` setting is not set; tokens will be signed with "
+                       "an insecure, static key")
+    secret_key = app.secret_key or 'NOT THAT SECRET'
+    app.tokenauth_serializer = JSONWebSignatureSerializer(secret_key)
+
+
+def claims_to_str(claims):
+    claims['v'] = TOKENAUTH_VERSION
+    return current_app.tokenauth_serializer.dumps(claims)
+
+
+def str_to_claims(token_str):
+    try:
+        claims = current_app.tokenauth_serializer.loads(token_str)
+    except BadData:
+        logger.warning("Got invalid signature in token %r", token_str)
+        return None
+
+    if claims.get('v') != TOKENAUTH_VERSION:
+        return None
+
+    return claims
